refactor(skills): split CLI into domain command modules (S1)
Split the 696-line src/cli/skills.ts into five domain modules under
src/cli/commands/skills/: index (dispatch), inventory (list/inspect +
human rendering helpers), draft (validate/publish-draft envelopes),
lifecycle (install/update/rollback/uninstall), and authoring
(create/enable/disable/copy-template). Each domain owns its own parse,
render, and error envelope. The old file is deleted with no pass-through
re-export; main.ts, doctor.ts, and two CLI test files import from the
new directory.
